requirements.txt
# ──────────────────────────────────────────────────────
# HybridMemoryNet — Brain Tumor Classification
# Python 3.9+  |  PyTorch 2.x  |  CUDA 11.8 / 12.x
# ──────────────────────────────────────────────────────

# Core
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0             # Swin-Tiny pretrained weights

# Data / utils
numpy>=1.23
Pillow>=9.0
PyYAML>=6.0
scikit-learn>=1.2
opencv-python>=4.7

# Visualisation
matplotlib>=3.7
tensorboard>=2.13

# Optional: faster data loading
# nvidia-dali-cuda110    # uncomment if you want DALI loader

# Dev / testing
tqdm>=4.65
